Ya funciona correctamente el calculo de energia y magnetizacion

+++ src/metropolis.h
@@ -3,7 +3,7 @@
 float metropolis(int *lattice, int n, float *energias, float energia, float J, float B, int *magnetizacion);
 int pick_site(int *lattice, int n);
 float flip(int *lattice, int n, float *energias, float energia, float J, float B, int *magnetizacion);
-float delta_energia(int *lattice, int n, int sitio);
+int delta_energia(int *lattice, int n, int sitio);
 float energia_total(int *lattice, int n, float J, float B);
 int magnetizacion_total(int *lattice, int n);
 #endif
